+++ src/main/java/uk/ac/gla/terrier/probos/master/webapp/ConfServlet.java
@@ -22,6 +22,7 @@ import java.util.List;
 import java.util.Map.Entry;
 
 import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -39,7 +40,7 @@ public class ConfServlet extends BaseServlet {
 	
 	ProbosApplicationMasterServiceImpl pams;
 	
-	public ConfServlet(String _uri, List<Entry<String,BaseServlet>> _servletNameSpace, PBSClient _pbsClient,
+	public ConfServlet(String _uri, List<Entry<String,HttpServlet>> _servletNameSpace, PBSClient _pbsClient,
 			ProbosApplicationMasterServiceImpl _pams)
 			throws Exception {
 		super(NAME, _uri, _servletNameSpace, _pbsClient);
@@ -51,7 +52,9 @@ public class ConfServlet extends BaseServlet {
 			HttpServletResponse resp, PrintStream ps) throws ServletException,
 			IOException {
 		
-		Configuration.dumpConfiguration(pams.getConf(), new OutputStreamWriter(ps));		
+		ps.println("<code class=\"json\">");
+		Configuration.dumpConfiguration(pams.getConf(), new OutputStreamWriter(ps));
+		ps.println("</code>");	
 	}
 
 }