+++ src/main/java/uk/ac/gla/terrier/probos/master/webapp/LogsServlet.java
@@ -23,11 +23,13 @@ import java.util.List;
 import java.util.Map.Entry;
 
 import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.hadoop.yarn.api.records.ContainerId;
 
+import uk.ac.gla.terrier.probos.Constants;
 import uk.ac.gla.terrier.probos.api.PBSClient;
 import uk.ac.gla.terrier.probos.common.BaseServlet;
 import uk.ac.gla.terrier.probos.master.ProbosApplicationMasterServiceImpl;
@@ -40,7 +42,7 @@ public class LogsServlet extends BaseServlet {
 	
 	ProbosApplicationMasterServiceImpl pams;
 	
-	public LogsServlet(String _uri, List<Entry<String,BaseServlet>> _servletNameSpace, PBSClient _pbsClient,
+	public LogsServlet(String _uri, List<Entry<String,HttpServlet>> _servletNameSpace, PBSClient _pbsClient,
 			ProbosApplicationMasterServiceImpl _pams)
 			throws Exception {
 		super(NAME, _uri, _servletNameSpace, _pbsClient);
@@ -57,10 +59,12 @@ public class LogsServlet extends BaseServlet {
 			throws ServletException, IOException
 	{
 		try{
-			ps.println("<p>");
+			
 			int jobId = pams.getJobId();
 			if (pams.isArray())
 			{
+				ps.println("<h1>Array Task Outputs</h1>");
+				ps.println("<p>");
 				ps.println("<ul>");
 				TIntObjectHashMap<ContainerId> array2cid = pams.getArrayIds();
 				for(int aid : array2cid.keys())
@@ -73,12 +77,24 @@ public class LogsServlet extends BaseServlet {
 						+"</li>");					
 				}
 				ps.println("</ul>");
+				ps.println("</p>");
 			} else {
+				ps.println("<h1>Job Outputs</h1>");
+				ps.println("<p>");
 				String stdout = new String( super.c.jobLog(jobId, -1, true, 0, true) );
 				String stderr = new String( super.c.jobLog(jobId, -1, false, 0, true) );				
 				ps.println("<a href=\""+stdout+"\">stdout</a>" + ", " + "<a href=\""+stderr+"\">stderr</a>");
+				ps.println("</p>");
 			}
-			ps.println("</p>");
+			
+			ps.println("<p>&nbsp;</p>");
+			ps.println("<h1>Master Outputs (for debugging "+Constants.PRODUCT_FRIENDLY_NAME+")</h1>");
+			ps.println("<p>");
+			String stdout = new String( super.c.jobLog(-1 * jobId, -1, true, 0, true) );
+			String stderr = new String( super.c.jobLog(-1 * jobId, -1, false, 0, true) );				
+			ps.println("<a href=\""+stdout+"\">stdout</a>" + ", " + "<a href=\""+stderr+"\">stderr</a>");
+			ps.println("</p>");		
+			
 		} catch (Exception e) {
 			throw new IOException(e);
 		}